move policy object and audit log definitions into their own file to keep audit.go clean
